Finish PlayersLoader decorator
Working on Players spec and then on to implementing the PlayersLoader decorator in the Database controller. The final step to getting the top level integration spec of loading the players to pass will be to complete the MVC loop and write the correct view code for the database #show page.