Finish Dataset and Players associations
Add   association to  and  association to , also add dataset_id index to  model. Dataset_id is being set in the  class of the ETL process. This is a seemingly 'dirty' way of doing this as I don't necessarily believe the Load class should have any knowledge of the Dataset per se. I'll have to investigate other options of setting the dataset id for each Player during their creation. Potentially if a Players Controller is created later on, this logic could be pushed their, but there is no need for a Players Controller at this point.

To-do:

Consider more tests surrounding the . Good and bad scenarios are needed.
Investigate other ways of assigning the .