Add initial specs for combinator class
Initially tried to use FactoryGirl for Combinator but it was more trouble that it was worth. The initialization tests for number of arguments is not needed now that the Combinator class is being initialized with a arguments hash. No point in checking for the bad path regarding arguments at this point. On to the business logic of forming lineup combinations.